move seq2seq ppl code into seq2seq.py
