+++ parlai/agents/seq2seq/seq2seq.py
@@ -15,8 +15,9 @@ import torch
 from torch.autograd import Variable
 from torch import optim
 import torch.nn as nn
+import torch.nn.functional as F
 
-from collections import deque
+from collections import deque, defaultdict
 
 import os
 import math
@@ -665,3 +666,63 @@ class Seq2seqAgent(Agent):
         """Use the metrics to decide when to adjust LR schedule."""
         if 'loss' in metrics_dict:
             self.scheduler.step(metrics_dict['loss'])
+
+
+class mydefaultdict(defaultdict):
+    """Custom defaultdict which overrides defaults requested by the get
+    function with the default factory.
+    """
+    def get(self, key, default=None):
+        # override default from "get" (like "__getitem__" already is)
+        return super().get(key, self.default_factory())
+
+
+class Seq2seqPerplexityAgent(Seq2seqAgent):
+    """Subclass for doing standardized perplexity evaluation.
+
+    This is designed to be used in conjunction with the PerplexityWorld at
+    parlai/scripts/eval_ppl.py. It uses the `next_word_probability` function
+    to calculate the probability of tokens one token at a time.
+    """
+
+    def __init__(self, opt, shared=None):
+        super().__init__(opt, shared)
+        self.prev_enc = None
+
+    def next_word_probability(self, partial_out):
+        """Return probability distribution over next words given an input and
+        partial true output. This is used to calculate the per-word perplexity.
+
+        Arguments:
+        observation -- input observation dict
+        partial_out -- list of previous "true" words
+
+        Returns a dict, where each key is a word and each value is a probability
+        score for that word. Unset keys assume a probability of zero.
+
+        e.g.
+        {'text': 'Run test program.'}, ['hello'] => {'world': 1.0}
+        """
+        obs = self.observation
+        obs['eval_labels'] = [' '.join(partial_out)]
+        batch = self.vectorize([obs])
+        if self.prev_enc is not None and batch[0].shape[1] != self.prev_enc[0].shape[1]:
+            self.prev_enc = None  # reset prev_enc
+
+        self.model.eval()
+        self.model.longest_label = 1  # no need to predict farther ahead
+        out = self.model(
+            batch[0], # xs
+            ys=(batch[1] if len(partial_out) > 0 else None),
+            prev_enc=self.prev_enc)
+        scores, self.prev_enc = out[1], out[-1]
+        # scores is bsz x seqlen x num_words, so select probs of current index
+        probs = F.softmax(scores.select(1, -1), dim=1).squeeze()
+        dist = mydefaultdict(lambda: 1e-7)  # minimum probability for any guess
+        for i in range(len(probs)):
+            try:
+                val = probs[i].item()
+            except AttributeError:
+                val = probs[i][0]
+            dist[self.dict[i]] = val
+        return dist
