+++ projects/convai2/baselines/seq2seq/eval_ppl.py
@@ -7,73 +7,14 @@
 This seq2seq model was trained on convai2:self.
 """
 from parlai.core.build_data import download_models
-from parlai.core.dict import DictionaryAgent
-from parlai.core.params import ParlaiParser, modelzoo_path
-from parlai.agents.seq2seq.seq2seq import Seq2seqAgent
-from projects.convai2.build_dict import build_dict
 from projects.convai2.eval_ppl import setup_args, eval_ppl
-import torch.nn.functional as F
 
-
-class Seq2seqEntry(Seq2seqAgent):
-    def __init__(self, opt, shared=None):
-        super().__init__(opt, shared)
-        if shared:
-            self.probs = shared['probs']
-        else:
-            # default minimum probability mass for all tokens
-            self.probs = {k: 1e-7 for k in build_dict().keys()}
-        self.prev_enc = None
-        self.prev_vec = None
-
-    def share(self):
-        shared = super().share()
-        shared['probs'] = self.probs.copy()
-        return shared
-
-    def next_word_probability(self, partial_out):
-        """Return probability distribution over next words given an input and
-        partial true output. This is used to calculate the per-word perplexity.
-
-        Arguments:
-        observation -- input observation dict
-        partial_out -- list of previous "true" words
-
-        Returns a dict, where each key is a word and each value is a probability
-        score for that word. Unset keys assume a probability of zero.
-
-        e.g.
-        {'text': 'Run test program.'}, ['hello'] => {'world': 1.0}
-        """
-        obs = self.observation
-        obs['eval_labels'] = [' '.join(partial_out)]
-        batch = self.vectorize([obs])
-        if self.prev_enc is not None and batch[0].shape[1] != self.prev_enc[0].shape[1]:
-            self.prev_enc = None  # reset prev_enc
-
-        self.model.eval()
-        self.model.longest_label = 1  # no need to predict farther ahead
-        out = self.model(
-            batch[0], # xs
-            ys=(batch[1] if len(partial_out) > 0 else None),
-            prev_enc=self.prev_enc)
-        scores, self.prev_enc = out[1], out[-1]
-        # scores is bsz x seqlen x num_words, so select probs of current index
-        probs = F.softmax(scores.select(1, -1), dim=1).squeeze()
-        dist = self.probs
-        for i in range(len(probs)):
-            try:
-                val = probs[i].item()
-            except AttributeError:
-                val = probs[i][0]
-            dist[self.dict[i]] = val
-        return dist
 
 
 if __name__ == '__main__':
     parser = setup_args()
     parser.set_params(
-        model='projects.convai2.baselines.seq2seq.eval_ppl:Seq2seqEntry',
+        model='parlai.agents.seq2seq.seq2seq:Seq2seqPerplexityAgent',
         model_file='models:convai2/seq2seq/convai2_self_seq2seq_model',
         dict_file='models:convai2/seq2seq/convai2_self_seq2seq_model.dict',
         dict_lower=True,
